fix(grammar,parser): restrict deferred connections to one at grammar level to fix issue with chain inside defer
