+++ e2e/order_dependend_with_arr_inport/main/main.neva
@@ -4,11 +4,11 @@ component Main(start) (stop) {
 		sub PortReducer<int> { reducer Sub<int> }
 	}
 	net {
-		:start -> (
-			1 -> sub:port[0],
-			2 -> sub:port[1],
-			3 -> sub:port[2]
-		)
+		:start -> [
+			(1 -> sub:port[0]),
+			(2 -> sub:port[1]),
+			(3 -> sub:port[2])
+		]
 		sub:res -> println:data
 		println:sig -> :stop
 	}
