+++ examples/4_math/0_add_nums_verbose/main.neva
@@ -5,10 +5,10 @@ component Main(start any) (stop any) {
 		println Println<int>
 	}
 	net {
-		:start -> (
-			1 -> sequencer:port[0],
-			2 -> sequencer:port[1]
-		)
+		:start -> [
+			(1 -> sequencer:port[0]),
+			(2 -> sequencer:port[1])
+		]
 		sequencer:seq -> adder:seq
 		adder:res -> println:data
 		println:sig -> :stop
