+++ examples/4_math/1_add_nums_with_bridge/main.neva
@@ -4,10 +4,10 @@ component Main(start any) (stop any) {
 		println Println<int> 
 	}
 	net {
-		:start -> (
-			1 -> adder:port[0],
-			2 -> adder:port[1]
-		)
+		:start -> [
+			(1 -> adder:port[0]),
+			(2 -> adder:port[1])
+		]
 		adder:res -> println:data
 		println:sig -> :stop
 	}
