+++ examples/6_struct_builder/verbose/main.neva
@@ -13,10 +13,10 @@ component {
     Main(start any) (stop any) {
         nodes { Println<User>, builder MyStructBuilder }
         net {
-            :start -> (
-                'John' -> builder:name,
-                32 -> builder:age
-            )
+            :start -> [
+                ('John' -> builder:name),
+                (32 -> builder:age)
+            ]
             builder:msg -> println:data
             println:sig -> :stop
         }
