+++ internal/compiler/parser/tests/happypath/025_component.neva
@@ -238,22 +238,22 @@ component {
 		net {
             $foo -> (foo:bar -> bar:baz)
 
-			$foo -> (foo:bar -> bar:baz,
-			foo:bar -> bar:baz
-			)
+			$foo -> [(foo:bar -> bar:baz),
+			(foo:bar -> bar:baz)
+            ]
 
-			$foo -> (foo:bar -> bar:baz,
-				foo:bar -> bar:baz
-			)
+			$foo -> [(foo:bar -> bar:baz),
+				(foo:bar -> bar:baz)
+            ]
 
-			$foo -> (
-				foo:bar -> bar:baz,
-				foo:bar -> bar:baz)
+			$foo -> [
+				(foo:bar -> bar:baz),
+				(foo:bar -> bar:baz)]
 
-			$foo -> (
-				foo:bar -> bar:baz,
-				foo:bar -> bar:baz
-			)
+			$foo -> [
+				(foo:bar -> bar:baz),
+				(foo:bar -> bar:baz)
+            ]
 		}
 	}
 }
